+++ frontend/src/scenes/session-recordings/player/sessionRecordingDataLogic.ts
@@ -23,21 +23,15 @@ import { FEATURE_FLAGS } from 'lib/constants'
 import { Dayjs, dayjs } from 'lib/dayjs'
 import { featureFlagLogic, FeatureFlagsSet } from 'lib/logic/featureFlagLogic'
 import { isObject } from 'lib/utils'
-import { chainToElements } from 'lib/utils/elements-chain'
 import { eventUsageLogic } from 'lib/utils/eventUsageLogic'
 import posthog from 'posthog-js'
 import { compressedEventWithTime } from 'posthog-js/lib/src/extensions/replay/sessionrecording'
 import { RecordingComment } from 'scenes/session-recordings/player/inspector/playerInspectorLogic'
+import { sessionRecordingEventDataLogic } from 'scenes/session-recordings/player/sessionRecordingEventDataLogic'
 import { teamLogic } from 'scenes/teamLogic'
 
-import { HogQLQuery, NodeKind } from '~/queries/schema'
-import { hogql } from '~/queries/utils'
 import {
-    AnyPropertyFilter,
     EncodedRecordingSnapshot,
-    PersonType,
-    PropertyFilterType,
-    PropertyOperator,
     RecordingEventsFilters,
     RecordingEventType,
     RecordingReportLoadTimes,
@@ -59,7 +53,6 @@ import type { sessionRecordingDataLogicType } from './sessionRecordingDataLogicT
 import { createSegments, mapSnapshotsToWindowId } from './utils/segmenter'
 
 const IS_TEST_MODE = process.env.NODE_ENV === 'test'
-const BUFFER_MS = 60000 // +- before and after start and end of a recording to query for.
 const DEFAULT_REALTIME_POLLING_MILLIS = 3000
 
 let postHogEEModule: PostHogEE
@@ -356,35 +349,6 @@ export interface SessionRecordingDataLogicProps {
     realTimePollingIntervalMilliseconds?: number
 }
 
-function makeEventsQuery(
-    person: PersonType | null,
-    distinctId: string | null,
-    start: Dayjs,
-    end: Dayjs,
-    properties: AnyPropertyFilter[]
-): Promise<unknown> {
-    return api.query({
-        kind: NodeKind.EventsQuery,
-        // NOTE: Be careful adding fields here. We want to keep the payload as small as possible to load all events quickly
-        select: [
-            'uuid',
-            'event',
-            'timestamp',
-            'elements_chain',
-            'properties.$window_id',
-            'properties.$current_url',
-            'properties.$event_type',
-        ],
-        orderBy: ['timestamp ASC'],
-        limit: 1000000,
-        personId: person ? String(person.id) : undefined,
-        after: start.subtract(BUFFER_MS, 'ms').format(),
-        before: end.add(BUFFER_MS, 'ms').format(),
-        properties: properties,
-        where: distinctId ? [`distinct_id = ('${distinctId}')`] : undefined,
-    })
-}
-
 async function processEncodedResponse(
     encodedResponse: (EncodedRecordingSnapshot | string)[],
     props: SessionRecordingDataLogicProps,
@@ -416,10 +380,18 @@ export const sessionRecordingDataLogic = kea<sessionRecordingDataLogicType>([
     path((key) => ['scenes', 'session-recordings', 'sessionRecordingDataLogic', key]),
     props({} as SessionRecordingDataLogicProps),
     key(({ sessionRecordingId }) => sessionRecordingId || 'no-session-recording-id'),
-    connect({
+    connect((props: SessionRecordingDataLogicProps) => ({
         logic: [eventUsageLogic],
-        values: [featureFlagLogic, ['featureFlags'], teamLogic, ['currentTeam']],
-    }),
+        values: [
+            featureFlagLogic,
+            ['featureFlags'],
+            teamLogic,
+            ['currentTeam'],
+            sessionRecordingEventDataLogic(props),
+            ['sessionEventsDataLoading', 'loadEventsSuccess', 'loadEventsFailure'],
+        ],
+        actions: [sessionRecordingEventDataLogic(props), ['loadEvents', 'loadFulLEventData']],
+    })),
     defaults({
         sessionPlayerMetaData: null as SessionRecordingType | null,
     }),
@@ -432,8 +404,6 @@ export const sessionRecordingDataLogic = kea<sessionRecordingDataLogicType>([
         loadSnapshotSources: true,
         loadNextSnapshotSource: true,
         loadSnapshotsForSource: (source: Pick<SessionRecordingSnapshotSource, 'source' | 'blob_key'>) => ({ source }),
-        loadEvents: true,
-        loadFullEventData: (event: RecordingEventType | RecordingEventType[]) => ({ event }),
         markViewed: (delay?: number) => ({ delay }),
         reportUsageIfFullyLoaded: true,
         persistRecording: true,
@@ -584,149 +554,6 @@ export const sessionRecordingDataLogic = kea<sessionRecordingDataLogicType>([
                 },
             },
         ],
-        sessionEventsData: [
-            null as null | RecordingEventType[],
-            {
-                loadEvents: async () => {
-                    if (!cache.eventsStartTime) {
-                        cache.eventsStartTime = performance.now()
-                    }
-
-                    const { start, end, person } = values.sessionPlayerData
-
-                    if (!person || !start || !end) {
-                        return null
-                    }
-
-                    const [sessionEvents, relatedEvents]: any[] = await Promise.all([
-                        // make one query for all events that are part of the session
-                        makeEventsQuery(null, null, start, end, [
-                            {
-                                key: '$session_id',
-                                value: [props.sessionRecordingId],
-                                operator: PropertyOperator.Exact,
-                                type: PropertyFilterType.Event,
-                            },
-                        ]),
-                        // make a second for all events from that person,
-                        // not marked as part of the session
-                        // but in the same time range
-                        // these are probably e.g. backend events for the session
-                        // but with no session id
-                        // since posthog-js must always add session id we can also
-                        // take advantage of lib being materialized and further filter
-                        makeEventsQuery(null, values.sessionPlayerMetaData?.distinct_id || null, start, end, [
-                            {
-                                key: '$session_id',
-                                value: '',
-                                operator: PropertyOperator.Exact,
-                                type: PropertyFilterType.Event,
-                            },
-                            {
-                                key: '$lib',
-                                value: ['web'],
-                                operator: PropertyOperator.IsNot,
-                                type: PropertyFilterType.Event,
-                            },
-                        ]),
-                    ])
-
-                    return [...sessionEvents.results, ...relatedEvents.results].map(
-                        (event: any): RecordingEventType => {
-                            const currentUrl = event[5]
-                            // We use the pathname to simplify the UI - we build it here instead of fetching it to keep data usage small
-                            let pathname: string | undefined
-                            try {
-                                pathname = event[5] ? new URL(event[5]).pathname : undefined
-                            } catch {
-                                pathname = undefined
-                            }
-
-                            return {
-                                id: event[0],
-                                event: event[1],
-                                timestamp: event[2],
-                                elements: chainToElements(event[3]),
-                                properties: {
-                                    $window_id: event[4],
-                                    $current_url: currentUrl,
-                                    $event_type: event[6],
-                                    $pathname: pathname,
-                                },
-                                playerTime: +dayjs(event[2]) - +start,
-                                fullyLoaded: false,
-                            }
-                        }
-                    )
-                },
-
-                loadFullEventData: async ({ event }) => {
-                    // box so we're always dealing with a list
-                    const events = Array.isArray(event) ? event : [event]
-
-                    let existingEvents = values.sessionEventsData?.filter((x) => events.some((e) => e.id === x.id))
-
-                    const allEventsAreFullyLoaded =
-                        existingEvents?.every((e) => e.fullyLoaded) && existingEvents.length === events.length
-                    if (!existingEvents || allEventsAreFullyLoaded) {
-                        return values.sessionEventsData
-                    }
-
-                    existingEvents = existingEvents.filter((e) => !e.fullyLoaded)
-                    const timestamps = existingEvents.map((ee) => dayjs(ee.timestamp).utc().valueOf())
-                    const eventNames = Array.from(new Set(existingEvents.map((ee) => ee.event)))
-                    const eventIds = existingEvents.map((ee) => ee.id)
-                    const earliestTimestamp = timestamps.reduce((a, b) => Math.min(a, b))
-                    const latestTimestamp = timestamps.reduce((a, b) => Math.max(a, b))
-                    try {
-                        const query: HogQLQuery = {
-                            kind: NodeKind.HogQLQuery,
-                            query: hogql`SELECT properties, uuid
-                                         FROM events
-                                         WHERE timestamp > ${(earliestTimestamp - 1000) / 1000}
-                                           AND timestamp < ${(latestTimestamp + 1000) / 1000}
-                                           AND event in ${eventNames}
-                                           AND uuid in ${eventIds}`,
-                        }
-                        const response = await api.query(query)
-                        if (response.error) {
-                            throw new Error(response.error)
-                        }
-
-                        for (const event of existingEvents) {
-                            const result = response.results.find((x: any) => {
-                                return x[1] === event.id
-                            })
-
-                            if (result) {
-                                event.properties = JSON.parse(result[0])
-                                event.fullyLoaded = true
-                            }
-                        }
-                    } catch (e) {
-                        // NOTE: This is not ideal but should happen so rarely that it is tolerable.
-                        existingEvents.forEach((e) => (e.fullyLoaded = true))
-                        captureException(e, {
-                            tags: { feature: 'session-recording-load-full-event-data' },
-                        })
-                    }
-
-                    // here we map the events list because we want the result to be a new instance to trigger downstream recalculation
-                    return !values.sessionEventsData
-                        ? values.sessionEventsData
-                        : values.sessionEventsData.map((x) => {
-                              const event = existingEvents?.find((ee) => ee.id === x.id)
-                              return event
-                                  ? ({
-                                        ...x,
-                                        properties: event.properties,
-                                        fullyLoaded: event.fullyLoaded,
-                                    } as RecordingEventType)
-                                  : x
-                          })
-                },
-            },
-        ],
     })),
     listeners(({ values, actions, cache, props }) => ({
         loadSnapshots: () => {
@@ -745,7 +572,7 @@ export const sessionRecordingDataLogic = kea<sessionRecordingDataLogicType>([
         },
         loadSnapshotSources: () => {
             // We only load events once we actually start loading the recording
-            actions.loadEvents()
+            actions.loadEvents(values.sessionPlayerData)
         },
         loadRecordingMetaSuccess: () => {
             cache.metadataLoadDuration = Math.round(performance.now() - cache.metaStartTime)
@@ -885,12 +712,6 @@ export const sessionRecordingDataLogic = kea<sessionRecordingDataLogicType>([
         },
     })),
     selectors(({ cache }) => ({
-        webVitalsEvents: [
-            (s) => [s.sessionEventsData],
-            (sessionEventsData): RecordingEventType[] =>
-                (sessionEventsData || []).filter((e) => e.event === '$web_vitals'),
-        ],
-
         windowIdForTimestamp: [
             (s) => [s.segments],
             (segments) =>
